Move JarRelocatorHelper to a private subclass of DependencyManager, fix bug with using Path instead of File
